Fix: Made welcome page homepage and created projects path for all projects page
