--- app/welcome/page.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-function page() {
-  return (
-    <div>
-      <p>
-        im chosing welcon
-      </p>
-    </div>
-  )
-}
-
-export default page
